+++ src/value.rs
@@ -1,5 +1,10 @@
-use std::collections::HashMap;
+use crate::error::{self, Result};
 use base64::Engine;
+#[cfg(feature = "binary")]
+use msgpack_simple::{Extension, MapElement, MsgPack};
+use snafu::{ensure, ResultExt};
+
+use std::collections::HashMap;
 
 #[derive(Debug, Clone)]
 pub enum Value {
@@ -40,7 +45,7 @@ impl Value {
     pub fn as_string(&self) -> Option<&String> {
         match self {
             Self::String(value, ..) => Some(value),
-            _ => None
+            _ => None,
         }
     }
 
@@ -49,7 +54,6 @@ impl Value {
             Self::Bytes(value, ..) => Some(value),
             _ => None,
         }
-
     }
 
     pub fn as_int(&self) -> Option<i64> {
@@ -69,7 +73,7 @@ impl Value {
     pub fn as_bool(&self) -> Option<bool> {
         match self {
             Self::Bool(value, ..) => Some(*value),
-            _ => None
+            _ => None,
         }
     }
 
@@ -81,10 +85,7 @@ impl Value {
     }
 
     fn is_null(&self) -> bool {
-        match self {
-            Self::Null(_) => true,
-            _ => false,
-        }
+        matches!(self, Self::Null(_))
     }
 
     pub(crate) fn set_label(&mut self, value: &str) {
@@ -105,16 +106,280 @@ impl Value {
     pub(crate) fn to_macro_string(&self) -> String {
         match self {
             Self::Table(..) => self.to_string(),
-            Self::Array(children, _) => children.iter().map(|x| x.to_macro_string()).collect::<Vec<String>>().join(", "),
-            Self::Bytes(bytes, _) => base64::engine::general_purpose::URL_SAFE_NO_PAD.encode(bytes.as_slice()),
+            Self::Array(children, _) => children
+                .iter()
+                .map(|x| x.to_macro_string())
+                .collect::<Vec<String>>()
+                .join(", "),
+            Self::Bytes(bytes, _) => {
+                base64::engine::general_purpose::URL_SAFE_NO_PAD.encode(bytes.as_slice())
+            }
             Self::Int(value, _) => value.to_string(),
-            Self::Float(value, _ ) => value.to_string(),
+            Self::Float(value, _) => value.to_string(),
             Self::Bool(value, _) => value.to_string(),
             Self::Label(value) => value.clone(),
             Self::Null(_) => "null".to_owned(),
             Self::Macro(..) => self.to_string(),
             Self::String(value, _) => value.clone(),
         }
+    }
+
+    #[cfg(feature = "binary")]
+    pub fn from_binary(entry: msgpack_simple::MsgPack) -> Result<Self> {
+        let ext = entry
+            .as_extension()
+            .context(error::MsgPackNotExpectedSnafu)?;
+        match ext.type_id {
+            100 => {
+                let (label, content) = Self::extract(&ext)?;
+                let content = content.as_map().context(error::MsgPackNotExpectedSnafu)?;
+                let mut table_content = HashMap::new();
+                for entry in content.iter() {
+                    let key = entry
+                        .key
+                        .clone()
+                        .as_string()
+                        .context(error::MsgPackNotExpectedSnafu)?;
+                    let value = Self::from_binary(entry.value.clone())?;
+                    table_content.insert(key, value);
+                }
+                Ok(Self::Table(table_content, label))
+            }
+            101 => {
+                let (label, content) = Self::extract(&ext)?;
+                let content = content.as_array().context(error::MsgPackNotExpectedSnafu)?;
+                let mut array_content = Vec::new();
+                for entry in content.iter() {
+                    array_content.push(Self::from_binary(entry.clone())?);
+                }
+                Ok(Self::Array(array_content, label))
+            }
+            102 => {
+                let (label, content) = Self::extract(&ext)?;
+                let value = content
+                    .as_string()
+                    .context(error::MsgPackNotExpectedSnafu)?;
+                Ok(Self::String(value, label))
+            }
+            103 => {
+                let (label, content) = Self::extract(&ext)?;
+                let value = content
+                    .as_binary()
+                    .context(error::MsgPackNotExpectedSnafu)?;
+                Ok(Self::Bytes(value, label))
+            }
+            104 => {
+                let (label, content) = Self::extract(&ext)?;
+                let value = content.as_int().context(error::MsgPackNotExpectedSnafu)?;
+                Ok(Self::Int(value, label))
+            }
+            105 => {
+                let (label, content) = Self::extract(&ext)?;
+                let value = content.as_float().context(error::MsgPackNotExpectedSnafu)?;
+                Ok(Self::Float(value, label))
+            }
+            106 => {
+                let (label, content) = Self::extract(&ext)?;
+                let value = content
+                    .as_boolean()
+                    .context(error::MsgPackNotExpectedSnafu)?;
+                Ok(Self::Bool(value, label))
+            }
+            107 => {
+                let label = MsgPack::parse(ext.value.clone().as_slice())
+                    .context(error::MsgPackEncodedSnafu)?;
+                let label = label.as_string().context(error::MsgPackNotExpectedSnafu)?;
+                Ok(Self::Label(label))
+            }
+            108 => {
+                let (label, content) = Self::extract(&ext)?;
+                ensure!(content.is_nil(), error::MsgPackUnsupportedSnafu);
+                Ok(Self::Null(label))
+            }
+            _ => Err(error::Error::MsgPackUnsupported),
+        }
+    }
+
+    #[cfg(feature = "binary")]
+    fn extract(entry: &Extension) -> Result<(Option<String>, MsgPack)> {
+        let table =
+            MsgPack::parse(entry.value.clone().as_slice()).context(error::MsgPackEncodedSnafu)?;
+        let table = table.as_map().context(error::MsgPackNotExpectedSnafu)?;
+        // There should only be two entries in this encoding
+        ensure!(table.len() == 2, error::MsgPackUnsupportedSnafu);
+        let (key0, value0, value1) = (
+            table[0]
+                .key
+                .clone()
+                .as_int()
+                .context(error::MsgPackNotExpectedSnafu)?,
+            table[0].value.clone(),
+            table[1].value.clone(),
+        );
+        let label = if key0 == 0 {
+            value0.clone()
+        } else {
+            value1.clone()
+        };
+        let label = if label.is_nil() {
+            None
+        } else {
+            Some(label.as_string().context(error::MsgPackNotExpectedSnafu)?)
+        };
+        let content = if key0 == 1 {
+            value0.clone()
+        } else {
+            value1.clone()
+        };
+        Ok((label, content))
+    }
+
+    #[cfg(feature = "binary")]
+    pub fn to_binary(&self) -> msgpack_simple::MsgPack {
+        let (type_id, value) = match self {
+            Self::Table(data, label) => (
+                100,
+                MsgPack::Map(vec![
+                    MapElement {
+                        key: MsgPack::Int(0),
+                        value: label
+                            .as_ref()
+                            .map(|x| MsgPack::String(x.clone()))
+                            .unwrap_or(MsgPack::Nil),
+                    },
+                    MapElement {
+                        key: MsgPack::Int(1),
+                        value: MsgPack::Map(
+                            data.iter()
+                                .map(|(k, v)| MapElement {
+                                    key: MsgPack::String(k.clone()),
+                                    value: v.to_binary(),
+                                })
+                                .collect(),
+                        ),
+                    },
+                ]),
+            ),
+            Self::Array(data, label) => (
+                101,
+                MsgPack::Map(vec![
+                    MapElement {
+                        key: MsgPack::Int(0),
+                        value: label
+                            .as_ref()
+                            .map(|x| MsgPack::String(x.clone()))
+                            .unwrap_or(MsgPack::Nil),
+                    },
+                    MapElement {
+                        key: MsgPack::Int(1),
+                        value: MsgPack::Array(data.iter().map(|x| x.to_binary()).collect()),
+                    },
+                ]),
+            ),
+            Self::String(data, label) => (
+                102,
+                MsgPack::Map(vec![
+                    MapElement {
+                        key: MsgPack::Int(0),
+                        value: label
+                            .as_ref()
+                            .map(|x| MsgPack::String(x.clone()))
+                            .unwrap_or(MsgPack::Nil),
+                    },
+                    MapElement {
+                        key: MsgPack::Int(1),
+                        value: MsgPack::String(data.clone()),
+                    },
+                ]),
+            ),
+            Self::Bytes(data, label) => (
+                103,
+                MsgPack::Map(vec![
+                    MapElement {
+                        key: MsgPack::Int(0),
+                        value: label
+                            .as_ref()
+                            .map(|x| MsgPack::String(x.clone()))
+                            .unwrap_or(MsgPack::Nil),
+                    },
+                    MapElement {
+                        key: MsgPack::Int(1),
+                        value: MsgPack::Binary(data.clone()),
+                    },
+                ]),
+            ),
+            Self::Int(data, label) => (
+                104,
+                MsgPack::Map(vec![
+                    MapElement {
+                        key: MsgPack::Int(0),
+                        value: label
+                            .as_ref()
+                            .map(|x| MsgPack::String(x.clone()))
+                            .unwrap_or(MsgPack::Nil),
+                    },
+                    MapElement {
+                        key: MsgPack::Int(1),
+                        value: MsgPack::Int(*data),
+                    },
+                ]),
+            ),
+            Self::Float(data, label) => (
+                105,
+                MsgPack::Map(vec![
+                    MapElement {
+                        key: MsgPack::Int(0),
+                        value: label
+                            .as_ref()
+                            .map(|x| MsgPack::String(x.clone()))
+                            .unwrap_or(MsgPack::Nil),
+                    },
+                    MapElement {
+                        key: MsgPack::Int(0),
+                        value: MsgPack::Float(*data),
+                    },
+                ]),
+            ),
+            Self::Bool(data, label) => (
+                106,
+                MsgPack::Map(vec![
+                    MapElement {
+                        key: MsgPack::Int(0),
+                        value: label
+                            .as_ref()
+                            .map(|x| MsgPack::String(x.clone()))
+                            .unwrap_or(MsgPack::Nil),
+                    },
+                    MapElement {
+                        key: MsgPack::Int(1),
+                        value: MsgPack::Boolean(*data),
+                    },
+                ]),
+            ),
+            Self::Label(data) => (107, MsgPack::String(data.clone())),
+            Self::Null(label) => (
+                108,
+                MsgPack::Map(vec![
+                    MapElement {
+                        key: MsgPack::Int(0),
+                        value: label
+                            .as_ref()
+                            .map(|x| MsgPack::String(x.clone()))
+                            .unwrap_or(MsgPack::Nil),
+                    },
+                    MapElement {
+                        key: MsgPack::Int(1),
+                        value: MsgPack::Nil,
+                    },
+                ]),
+            ),
+            _ => return MsgPack::Nil,
+        };
+
+        MsgPack::Extension(Extension {
+            type_id,
+            value: value.encode(),
+        })
     }
 }
 
@@ -142,7 +407,15 @@ impl ToString for Value {
                 if let Some(label) = label {
                     s += format!("!{} ", label).as_str();
                 }
-                s += format!("[{}]", array.iter().map(|x| x.to_string()).collect::<Vec<String>>().join(", ")).as_str();
+                s += format!(
+                    "[{}]",
+                    array
+                        .iter()
+                        .map(|x| x.to_string())
+                        .collect::<Vec<String>>()
+                        .join(", ")
+                )
+                .as_str();
                 s
             }
             Self::String(string, label) => {
@@ -158,7 +431,11 @@ impl ToString for Value {
                 if let Some(label) = label {
                     s += format!("!{} ", label).as_str();
                 }
-                let m = if *is_string { format!("m!'{}'", string) } else { format!("m!{}", string) };
+                let m = if *is_string {
+                    format!("m!'{}'", string)
+                } else {
+                    format!("m!{}", string)
+                };
                 s += m.as_str();
                 s
             }
@@ -167,7 +444,8 @@ impl ToString for Value {
                 if let Some(label) = label {
                     s += format!("!{} ", label).as_str();
                 }
-                let encoded = base64::engine::general_purpose::URL_SAFE_NO_PAD.encode(bytes.as_slice());
+                let encoded =
+                    base64::engine::general_purpose::URL_SAFE_NO_PAD.encode(bytes.as_slice());
                 s += format!("b'{}'", encoded).as_str();
                 s
             }