+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+## 0.2.0 - Binary encoding support
+
+Added feature 'binary' that implements a conversion layer
+to encode BarkML into a binary representation through the use
+of `MessagePack`
+
 ## 0.1.0 - Initial Release
+
 Initial release of barkml.